requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
openai==1.3.7
whisper==1.1.10
torch==2.1.1
transformers==4.36.0
pydub==0.25.1
ffmpeg-python==0.2.0
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2 